Remove "mask" and "start" from ISection
"mask" and "start" are only needed internally, and adding them to
ISection means that they are required when creating a new Bitmask. This
is both unnecessary and meaningless, as "mask" and "start" are generated
in the constructor.

This requires creating a private "_sections" variable on Bitmask for
internal use, with the type of "BitmaskInternalSection", and making
"sections" a getter which simply returns "_sections" as the type
"ISection".

Note that these changes are only meaningful for TypeScript code, as
there are no functional changes.